fix(android): 修复Dashboard和Analytics页面JSON解析错误
修复了两个"JSON document was not fully consumed"错误：

1. **Dashboard统计页**：
   - 修复 DashboardStats.todayWorkTime 类型从 Long 改为 Int
   - API返回 Int 类型，之前 Kotlin 模型使用 Long 导致类型不匹配
   - 更新 UI 层添加 .toLong() 转换保持兼容性
   - 移除 DetailRepository 中不必要的 .toInt() 调用

2. **Analytics统计页**：
   - 修复 weekly-stats API 响应处理逻辑
   - 该 API 直接返回数据对象，无 {success, data, error} 包装层
   - 添加注释说明 API 响应格式以避免未来混淆
   - 恢复正确的响应解析：Response<WeeklyStatsResponse>

相关文件：
- Dashboard.kt: todayWorkTime Long → Int
- DashboardScreen.kt: 添加类型转换
- DetailRepository.kt: 移除冗余转换
- AnalyticsApi.kt: 添加 API 响应格式注释
- AnalyticsRepository.kt: 修复 weekly-stats 解析
- NetworkModule.kt: 添加 Gson 配置说明

+++ android-app/app/src/main/java/com/aiproj/mobile/data/models/Dashboard.kt
@@ -13,7 +13,7 @@ data class DashboardStats(
     val todayTasksTotal: Int = 0,
 
     @SerializedName("today_work_time")
-    val todayWorkTime: Long = 0, // 分钟
+    val todayWorkTime: Int = 0, // 分钟
 
     @SerializedName("active_projects")
     val activeProjects: Int = 0,
